+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+ - **Breaking**: Remove `store_locks` module, `[Async]ReadableWritableStorageTraits::mutex()`, and `new_with_locks` constructors from stores
+   - `DefaultStoreLocks` could result in a deadlock
+   - It is now the responsibility of zarrs consumers to ensure that:
+     - `Array::store_chunk_subset` is not called concurrently on the same chunk
+     - `Array::store_array_subset` is not called concurrently on regions sharing any chunks
+   - This may be revisited in a future release with an appropriate multi-chunk locking strategy
+
+## [0.13.3] - 2024-05-16
+
+*This release was yanked and changes reverted.*
+
 ## [0.13.2] - 2024-05-08
 
 ### Changed